@@ -1,6 +1,7 @@
 import assert = require('assert');
 import sinon = require('sinon');
-import fs = require('fs-promise');
+import * as fs from '../../src/workspace/readDirAsync';
+import path = require('path');
 
 import jsonFile = require('../../src/workspace/jsonFile');
 import { ProjectSettings } from '../../src/mavensmate/projectSettings';
@@ -35,11 +36,11 @@ suite('projectList', () => {
         readDirStub.withArgs('missingWorkspace').returns(Promise.reject('missingWorkspace is missing as intended'));
         
         jsonFileStub = sinon.stub(jsonFile, 'open');
-        jsonFileStub.withArgs('workspace1/project1/config/.settings').returns(testSettings);
-        jsonFileStub.withArgs('workspace1/project2/config/.settings').returns(testSettings);
-        jsonFileStub.withArgs('workspace2/project1/config/.settings').returns(testSettings);
-        jsonFileStub.withArgs('workspace2/project3/config/.settings').returns(testSettings);
-        jsonFileStub.withArgs('workspace2/doesNotExist/config/.settings').returns(null);
+        jsonFileStub.withArgs(path.normalize('workspace1/project1/config/.settings')).returns(testSettings);
+        jsonFileStub.withArgs(path.normalize('workspace1/project2/config/.settings')).returns(testSettings);
+        jsonFileStub.withArgs(path.normalize('workspace2/project1/config/.settings')).returns(testSettings);
+        jsonFileStub.withArgs(path.normalize('workspace2/project3/config/.settings')).returns(testSettings);
+        jsonFileStub.withArgs(path.normalize('workspace2/doesNotExist/config/.settings')).returns(null);
     });
     
     teardown(() => {
@@ -51,10 +52,10 @@ suite('projectList', () => {
     test('gets the 4 actual projects', (testDone) => {
         projectList.promiseList().then((projects) => {
                 assert.equal(projects.length, 4);
-                assertIsProject(projects[0], 'project1', 'workspace1/project1', 'workspace1');
-                assertIsProject(projects[1], 'project2', 'workspace1/project2', 'workspace1');
-                assertIsProject(projects[2], 'project1', 'workspace2/project1', 'workspace2');
-                assertIsProject(projects[3], 'project3', 'workspace2/project3', 'workspace2');
+                assertIsProject(projects[0], 'project1', path.normalize('workspace1/project1'), 'workspace1');
+                assertIsProject(projects[1], 'project2', path.normalize('workspace1/project2'), 'workspace1');
+                assertIsProject(projects[2], 'project1', path.normalize('workspace2/project1'), 'workspace2');
+                assertIsProject(projects[3], 'project3', path.normalize('workspace2/project3'), 'workspace2');
             })
             .done(testDone, console.error);
     });
